+++ CHANGELOG.md
@@ -21,7 +21,7 @@ The `identityEndpoint` prop has been replaced with a required `driver` prop.
 For applications that use JSON:API - you must import the JsonApi driver and use it as an argument to the provider:
 
 ```jsx
-import JsonApi from '@oris/auth/drivers/JsonApi';
+import JsonApi from '@oris/auth/dist/drivers/JsonApi';
 
 function App() {
     return (
@@ -46,7 +46,7 @@ By default, the `JsonApi` driver will use `/app-name/api/user` and `/app-name/ap
 A `GraphQL` driver is included for applications using [ORIS\GraphQL](https://code.osu.edu/oris/graphql).
 
 ```jsx
-import GraphQL from '@oris/auth/drivers/GraphQL';
+import GraphQL from '@oris/auth/dist/drivers/GraphQL';
 
 function App() {
     return (