+++ CHANGELOG.md
@@ -60,7 +60,8 @@ function App() {
 __New Features:__
 
 * Added support for GraphQL
-* Added `<Can>` component to check against permissions and policies.
+* Added `can(action: string, on?: IHasPolicies)` to the `useIdentity` hook to test against user permissions and contextual policies.
+* Added `<Can>` component to wrap around the aforementioned function.
     * Note that policies are **not** supported for the logged in user if using a `JsonApi` driver in AuthProvider. 
     * `JsonApi` does support usage for permission checks: `<Can do="foo.some-action">`
 * Added `<SimpleSignOutButton />` if you want a sign out button for an app to meet audit compliance without having to use the full `<Profile>` component.